requirements.txt
python-decouple==3.0
PyYAML==3.12
requests==2.12.4
tablib==0.11.3
